Dragalia: Prevent unintended healing condition
While the Dragon Shift is active, the player could cancel the mount and directly re-enter the Dragon Shift if the gauge has enough charge left. Each time the Dragon Shift is activated the player heals, creating an unintended heal loop. Now, when the Dragon Shift is cancelled by any means, the gauge is also reset. This successfully prevents the healing exploit.